Fix logic of showing Terms Confirm Dialog (#330, #331)
