+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/backup/BackupInfoFragment.kt
@@ -7,10 +7,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.ui.dialogs.BottomConfirmAlert
 import kotlinx.android.synthetic.main.fragment_backup_words_info.*
 
-class BackupInfoFragment : Fragment(), BottomConfirmAlert.Listener {
+class BackupInfoFragment : Fragment() {
     private lateinit var viewModel: BackupViewModel
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -27,19 +26,10 @@ class BackupInfoFragment : Fragment(), BottomConfirmAlert.Listener {
         buttonBackup.setOnClickListener { viewModel.delegate.showWordsDidClick() }
 
         buttonLater.setOnClickListener {
-            activity?.let {
-                val confirmationList = mutableListOf(
-                        R.string.Backup_Confirmation_Understand,
-                        R.string.Backup_Confirmation_DeleteAppWarn,
-                        R.string.Backup_Confirmation_LockAppWarn
-                )
-                BottomConfirmAlert.show(it, confirmationList, this)
-            }
+            viewModel.delegate.onLaterClick()
         }
+
     }
 
-    override fun onConfirmationSuccess() {
-        viewModel.delegate.onTermsConfirm()
-    }
 }
