+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/restore/RestoreInteractor.kt
@@ -1,10 +1,12 @@
 package io.horizontalsystems.bankwallet.modules.restore
 
 import io.horizontalsystems.bankwallet.core.IKeyStoreSafeExecute
+import io.horizontalsystems.bankwallet.core.ILocalStorage
 import io.horizontalsystems.bankwallet.core.managers.WordsManager
 
 class RestoreInteractor(
         private val wordsManager: WordsManager,
+        private val localStorage: ILocalStorage,
         private val keystoreSafeExecute: IKeyStoreSafeExecute) : RestoreModule.IInteractor {
 
     var delegate: RestoreModule.IInteractorDelegate? = null
@@ -14,6 +16,7 @@ class RestoreInteractor(
                 action = Runnable { wordsManager.restore(words) },
                 onSuccess = Runnable {
                     wordsManager.isBackedUp = true
+                    localStorage.iUnderstand = true
                     delegate?.didRestore()
                 },
                 onFailure = Runnable { delegate?.didFailToRestore() }
